+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this extension will be documented in this file.
 
 ## Unreleased
 
+### Fixed
+
+- Fixed an issue where the extension would fail to sync CCloud authentication related user settings
+  to the sidecar process when updating the extension.
+
 ## 1.0.1
 
 ### Fixed