+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this extension will be documented in this file.
 ### Fixed
 
 - Improved resolving newly created direct connection connectivity state.
+- Fixed an issue where the extension would fail to sync CCloud authentication related user settings
+  to the sidecar process when updating the extension.
 
 ## 1.0.1